completed changes that nikhil requested
